Finished FollowNodes script and updated playermovement script
Updated the follownodes script to fully work. Includes speed, pause at
nodes, and the ability to mirror the path when the last node is reached.
Updated player movement to apply the changes in motion in the fixed
update loop instead of normal update loop. The goal is to try and not
have as many collision glitches where colliders meet at the same point.