Readded players and controls, updated camera, General Polish
Multiplayer controls and prefabs were missing even after main branch
merge. Readded controls (should probably be updated) to be able to test
players.
Added new level "Basic Square" for testing things.
Updated Camera Follow script. It was allowing players to march off the
screen when 3 players were going the opposite way. Now it is based on
the furthest distance (on x or y axis) and zooms in and out based on the
widest range of x and y. This "should" mean it's now impossible for
players to walk off the screen since it's only based on min/max values
instead of averages.
Updated/added some comments to Camera Follow.
Made prefabs for the 4 players as well as a prefab including all 4
players to just be dragged into the scene.
Added/updated GUI for the 4 players. All monies are now individually
displayed rather then only updating the money for whoever picked
something up last.