Almost all of this was to get the money gui text updating properly, and to clean up a bit of code. If we need to go back to GUIHandler, that's fine, but I didn't see the reason for it. May make it an array in gamemanager instead of having to find the objects on load. Would only make it milliseconds faster on load though, if even.
Update CameraFollow Script to use an array with preset size to allow adding objects in code instead of reference in the editor.
In GameManager, Make PlayerCount a slider... for reasons... cause... yeah...
In GameManager, Made Player#sprite an array instead of individual objects to make cycling through them easier
In GameManager, Changed players from list to a simple array and hid it in the editor.
In GameManager, Cleaned up a lot of the code for adding players Removed the extra for loop and made the objectstofollow variable to send to the camera an array instead of a list.
In GameManager, removed getting canvas and playerstatus, as they weren't used.
In GameManager, removed GUIHandler as it wasn't working and was complicating things (I most likely just didn't understand what Wes was doing.)
In GameManager, Updated process from getting type of camerafollow to just getting the component from the main camera. Since there is always just one camera, there's no reason to use a find of type instead of direct reference.
In Status, Added playerID to make self reference easier.
In Status, Added moneyText variable to reference gui text for money updates easier
In Status, Added a find for the moneyText object using individual player ids.
In Status, Removed GUIHandler reference as it was never getting set in GameManager (even before removing it) since the GUIHandeler script was never attached to anything in any scenes (that I could find)