Added player creation in game maker to InitGame Function instead of awake. Added boolean to see if we need to create characters on it. This will make it easier to set up characters at a character select screen as well as have "cutscenes" be loadable without player objects in it.
Added player attacking by having a "weapon" object attached to player movement script.
PlayerMovement was updated to rotate and position the weapon when turning and such. Should probably be updated to call a script on the weapon so it rotates itself to allow different sizes and positions based on different weapons.
Added idle (blank image) and attack animations for a sword. They enable/disable the collider as well, so we don't have to code it. Weapon just needs damage dealer script set to enemies.
Don't think it's in this update, but I updated the layer detection in damage dealer to work properly. Apparently it was only seeing the player layer for some reason. Should work now.
Updated all 4 player prefabs to have weapon, with proper linking.